ci: try different version tag to avoid collisions
I have no idea why I'm getting "Bad Credentials" from the GitHub API.
The script block in the workflow works when I run it as a script locally
and I've tried granting the workflow write-all. I'm wondering now
whether the existing release is inaccessible to the token but a
non-existent release would be allowed. Seems like a longshot but I don't
know what else to try.